Add a fun to correct getCurrentLiveOffset value
After a stream loses connection with the serve the stream will not end
as there is a twitch feature to keep the stream open for sometime to see if it recuperates
so the stream doesn't loses views.

But when that happens the stream TS tag EXT-X-PROGRAM-DATE-TIME get out of sink
causing the LiveOffset to be wrong using a simple check we can prevent it.

Signed-off-by: fgl27 <fglfgl27@gmail.com>